+++ src/domain/transaction/transaction.go
@@ -2,10 +2,9 @@ package transaction
 
 import "time"
 
-// Transação representa uma única transação financeira
 type Transaction struct {
 	ID        int       `json:"id"`
-	Amount    float64   `json:"amount"`
+	Amount    int   `json:"amount"`
 	Type      string    `json:"type"` // "credit" ou "debit"
 	CreatedAt time.Time `json:"CreatedAt"`
 	CreatedBy string    `json:"CreatedBy"`
